TestCases/requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
async-generator==1.10
attrs==21.4.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
chromedriver-autoinstaller==0.3.1
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.2
decorator==5.1.1
docopt==0.6.2
et-xmlfile==1.1.0
execnet==1.9.0
h11==0.13.0
idna==3.3
iniconfig==1.1.1
mysql-connector-python==8.0.28
numpy==1.22.3
oauthlib==3.2.0
openpyxl==3.0.9
outcome==1.1.0
packaging==21.3
pandas==1.4.2
paramiko==2.10.3
pipreqs==0.4.11
pluggy==1.0.0
prettytable==3.2.0
protobuf==3.19.3
py==1.11.0
pycparser==2.21
PyMySQL==1.0.2
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.8
pytest==7.1.1
pytest-dependency==0.5.1
pytest-forked==1.4.0
pytest-html==3.1.1
pytest-json-report==1.5.0
pytest-metadata==2.0.1
pytest-order==1.0.1
pytest-ordering==0.6
pytest-parallel==0.1.1
pytest-xdist==2.5.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
selenium==4.0.0
selenium-webdriver-extender==0.0.0
self==2020.12.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sshtunnel==0.4.0
tblib==1.7.0
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.9
wcwidth==0.2.5
webdriver-manager==3.5.4
wsproto==1.1.0
XlsxWriter==3.0.3
yarg==0.1.9
cv2_extras==0.5.4
